+++ packages/expo-video/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - [Android][iOS] Added support for Audio Track feature. You can now set the audio track using `player.audioTrack` and list available audio tracks using `player.availableAudioTracks`. ([#36207](https://github.com/expo/expo/pull/36207) by [@HADeveloper](https://github.com/HADeveloper))
+- [Android][iOS] Add `keepScreenOnWhilePlaying` property to the player.
 
 ### 🐛 Bug fixes
 
@@ -19,6 +20,8 @@
 - [iOS] Fix player getting stuck in `loading` state for null sources. ([#37183](https://github.com/expo/expo/pull/37183) by [@behenate](https://github.com/behenate))
 
 ### 💡 Others
+
+- [Android] Keep the screen on while playback is running by default.
 
 ## 2.1.9 — 2025-05-08